MAIN golang structure
And it just works😉
