+++ ToMovies.go
@@ -1,13 +1,9 @@
 package main
 
-import (
-	. "github.com/OmniMir/AutoGo/libs"
-)
-
 const (
 	premieres = "www.kinopoisk.ru/premiere/ru/"
 )
 
-func main() {
+func ToMovies() {
 	StartChrome(premieres)
 }
