+++ Variables.go
@@ -1,0 +1,5 @@
+package main
+
+const (
+Space  = " "
+)
