--- libs/variables.go
@@ -1,8 +1,0 @@
-package libs
-
-const (
-Space  = " "
-)
-
-func main () {
-}
